requirements.txt
torch
torchvision
ftfy
regex
tqdm
clip @ git+https://github.com/openai/CLIP.git
Pillow
gradio>=4.0
